:zap: Add versioning
